fix doc tests

+++ src/schemes/boundary/system.jl
@@ -130,7 +130,7 @@ is_moving(t) = t < 1.5
 movement = BoundaryMovement(movement_function, is_moving)
 
 # output
-BoundaryMovement{typeof(movement_function), typeof(is_moving)}(movement_function, is_moving, Int[])
+BoundaryMovement{typeof(movement_function), typeof(is_moving)}(movement_function, is_moving, Int64[])
 ```
 """
 struct BoundaryMovement{MF, IM}
